Continued with different algorithms and worked with performance tests.
Finished the scoring method in Optimized Algo, corrected an error in
the score card (had to do with bonus) and started working with a first
look approach algorithm that is to play better than the human algo..
Also started making performance tests! One test for the Human algo and
one for the BruteAlgo. Gives a good understanding for the results. Next
is to work on printing these results to an excel file.